+++ src/app/components/sidenav.component.ts
@@ -1,7 +1,10 @@
-import { Component, ViewEncapsulation } from '@angular/core';
+import { Component, ViewEncapsulation, OnInit } from '@angular/core';
 import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { map, filter } from 'rxjs/operators';
+import { Store } from '@ngrx/store';
+import { AppState } from '../store/app.reducers';
+import { UiState } from '../store/reducers/ui.reducer';
 
 @Component({
   selector: 'app-sidenav',
@@ -17,7 +20,6 @@ import { map } from 'rxjs/operators';
 
         <mat-nav-list>
           <a mat-list-item routerLink="/">Users</a>
-          <a mat-list-item routerLink="user/0">Profile</a>
         </mat-nav-list>
       </mat-sidenav>
 
@@ -32,14 +34,11 @@ import { map } from 'rxjs/operators';
             <mat-icon aria-label="Side nav toggle icon">menu</mat-icon>
           </button>
 
-          <span>ngrx-effects</span>
-
-          <div class="sidenav-search">
-            <mat-form-field appearance="outline" floatLabel="never">
-              <input (keyup)="search($event)" matInput placeholder="User ID">
-            </mat-form-field>
-          </div>
+          <span>Effects example</span>
         </mat-toolbar>
+
+        <mat-progress-bar mode="query"
+          *ngIf="(uiState$ | async)?.appIsLoading;"></mat-progress-bar>
 
         <main class="main">
           <router-outlet></router-outlet>
@@ -77,11 +76,23 @@ import { map } from 'rxjs/operators';
     }
   `]
 })
-export class SidenavComponent {
-  isHandset$: Observable<boolean> = this.breakpointObserver.observe(Breakpoints.Handset)
+export class SidenavComponent implements OnInit {
+  public uiState$: Observable<UiState>;
+  public isHandset$: Observable<boolean> = this.breakpointObserver.observe(Breakpoints.Handset)
     .pipe(map(result => result.matches));
 
-  constructor(private breakpointObserver: BreakpointObserver) {}
+  constructor(
+    private breakpointObserver: BreakpointObserver,
+    private store: Store<AppState>
+  ) {
+
+  }
+
+  ngOnInit(): void {
+    setTimeout(() => {
+      this.uiState$ = this.store.select('ui');
+    });
+  }
 
   public search(event: any) {
     console.log(event.target.value);